Stub out some of the needed files, just need one more gen_server with some simple workers which will hold tcp connections to actually transfer files.
Updated script in priv/ and the rebar.config.script to be able to start a cluster with a configurable number of nodes, node names are just passed through
via the vm.args file and the start_cluster.sh script.

Pool will be a simple worker pool based off of poolboy, the checking out of a worker will be blocking which will help limit the number of files that
can be sent through the system at once.

es3_log will be a simple disk_log wrapper if i get time to log some debug messages to disk.